apps/agent-host/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
httpx>=0.26.0
openai>=1.12.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
pyyaml>=6.0
python-dotenv>=1.0.0
pycryptodome>=3.20.0
watchfiles>=0.21.0
lark-oapi>=1.5.0
discord.py>=2.4.0
APScheduler>=3.10.4
chromadb>=0.5.0
